+++ src/ch/systemsx/sybit/crkwebui/client/gui/renderers/MethodCellRenderer.java
@@ -37,7 +37,7 @@ public class MethodCellRenderer implements GridCellRenderer<BaseModel>
 		{
 			String tooltipText = null;
 			
-			int interfaceId = model.get("id");
+			int interfaceId = (Integer)model.get("id");
 			
 			InterfaceItem interfaceItem = mainController.getPdbScoreItem().getInterfaceItem(interfaceId - 1);
